Changed firebase structure, added component and page for transactions, delete function

+++ src/pages/CategoriesList/CategoriesList.tsx
@@ -2,7 +2,7 @@ import { useDispatch, useSelector } from "react-redux";
 import CategoryCard from "../../components/CategoryCard/CategoryCard";
 import type { AppDispatch, RootState } from "../../app/store";
 import { useEffect } from "react";
-import { deleteCategorie, fetchCategories } from "../../app/categorysSlice";
+import { deleteCategorie, fetchCategories } from "../../app/categoriesSlice";
 import Spinner from "../../components/Spinner/Spinner";
 import { NavLink } from "react-router-dom";
 
@@ -22,17 +22,29 @@ const CategoriesList = () => {
   }
 
   const handleDelete = async (id: string) => {
-   await dispatch(deleteCategorie(id))
-};
+    const isConfirmed = confirm("Delete category?");
+
+    if (!isConfirmed) return;
+
+    await dispatch(deleteCategorie(id));
+  };
 
   return (
     <>
       <div className="d-flex justify-content-between align-items-center mb-3 mt-5">
         <h5>Categories</h5>
-        <NavLink to={'/add-category'} className="btn btn-outline-primary">Add categorie</NavLink>
+        <NavLink to={"/add-category"} className="btn btn-outline-primary">
+          Add categorie
+        </NavLink>
       </div>
       {categories.map((cat) => (
-        <CategoryCard key={cat.id} name={cat.name} type={cat.type} onDelete={() => handleDelete(cat.id)} id={cat.id} />
+        <CategoryCard
+          key={cat.id}
+          name={cat.name}
+          type={cat.type}
+          onDelete={() => handleDelete(cat.id)}
+          id={cat.id}
+        />
       ))}
     </>
   );
